// src/generated/models/ConfigSLA.ts
export type ConfigSLAModel = runtime.Types.Result.DefaultSelection<Prisma.$ConfigSLAPayload>
// ...
export type $ConfigSLAPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "ConfigSLA"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    prioridad: $Enums.Prioridad
    horasLimite: number
    updatedAt: Date
  }, ExtArgs["result"]["configSLA"]>
  composites: {}
}